combined changes from multiple commits together
